+++ src/Function/JsonContainsPath.php
@@ -37,7 +37,7 @@ class JsonContainsPath extends AbstractJsonSearch
             return \sprintf('%s(%s, %s, %s)', static::FUNCTION_NAME, $jsonDocumentSql, $modeSql, \implode(', ', $walkedPaths));
         }
 
-        throw new Exception(\sprintf('method `%s` is not supported', static::FUNCTION_NAME));
+        throw new Exception(\sprintf('function `%s` is not supported', static::FUNCTION_NAME));
     }
 
     public function parse(Parser $parser): void