+++ src/Function/JsonExtract.php
@@ -37,7 +37,7 @@ class JsonExtract extends FunctionNode
             return \sprintf('%s(%s, %s)', static::FUNCTION_NAME, $jsonDocumentSql, \implode(', ', $walkedPaths));
         }
 
-        throw new Exception(\sprintf('method `%s` is not supported', static::FUNCTION_NAME));
+        throw new Exception(\sprintf('function `%s` is not supported', static::FUNCTION_NAME));
     }
 
     public function parse(Parser $parser): void