+++ src/Function/JsonSearch.php
@@ -47,7 +47,7 @@ class JsonSearch extends AbstractJsonSearch
             return \sprintf('%s(%s, %s, %s)', static::FUNCTION_NAME, $jsonDocumentSql, $modeSql, $searchArgsSql);
         }
 
-        throw new Exception(\sprintf('method `%s` is not supported', static::FUNCTION_NAME));
+        throw new Exception(\sprintf('function `%s` is not supported', static::FUNCTION_NAME));
     }
 
     public function parse(Parser $parser): void