+++ src/Function/JsonUnquote.php
@@ -30,7 +30,7 @@ class JsonUnquote extends FunctionNode
             return \sprintf('%s(%s)', static::FUNCTION_NAME, $jsonValueSql);
         }
 
-        throw new Exception(\sprintf('method `%s` is not supported', static::FUNCTION_NAME));
+        throw new Exception(\sprintf('function `%s` is not supported', static::FUNCTION_NAME));
     }
 
     public function parse(Parser $parser): void